requirements.txt
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.4
Faker==13.3.0
faker-vehicle==0.2.0
